separated App into custom hooks
